- fixed some smarty.const.PHP_SELF to the correct smarty.server.PHP_SELF which came in with commit 2132 - (also brought in the change to wfb.image_resize.php which was missing from my last commit) - relates to issue #353835

+++ webservices/wfb.image_resize.php
@@ -73,7 +73,9 @@ $is_url = false;
 if (preg_match('/^https?:\/\//',$image_dir.$image)) {
   // make a temporary local copy
   $is_url = true;
-  file_put_contents(APP_TEMP_DIR.$image, file_get_contents($image_dir.$image));
+  if(!is_file(APP_TEMP_DIR.$image)) {
+		file_put_contents(APP_TEMP_DIR.$image, file_get_contents($image_dir.$image));
+  }
   $image_dir = APP_TEMP_DIR;
 }
